rename worker directory to codalabworker
