+++ codalab/client/json_api_client.py
@@ -9,8 +9,8 @@ from codalab.common import (
     precondition,
     UsageError,
 )
-from worker.rest_client import RestClient, RestClientException
-from worker.file_util import un_gzip_stream
+from codalabworker.rest_client import RestClient, RestClientException
+from codalabworker.file_util import un_gzip_stream
 
 
 def wrap_exception(message):
