+++ codalab/lib/formatting.py
@@ -7,7 +7,7 @@ import json
 import shlex
 import pipes
 
-from worker import formatting as worker_formatting
+from codalabworker import formatting as worker_formatting
 
 
 NONE_PLACEHOLDER = '<none>'
