+++ codalabworker/docker_image_manager.py
