+++ codalabworker/worker.sh
