+++ test-cli.py
@@ -245,7 +245,7 @@ def temp_instance():
 
         # Start worker after servers are up
         worker_proc = subprocess.Popen([os.path.join(base_path, 'venv/bin/python'),
-                                        os.path.join(base_path, 'worker/main.py'),
+                                        os.path.join(base_path, 'codalabworker/main.py'),
                                         '--server=' + remote_host,
                                         '--slots=1',
                                         '--work-dir=' + remote_worker_scratch,
