+++ tests/worker/file_util_test.py
@@ -2,7 +2,7 @@ import os
 import tempfile
 import unittest
 
-from worker.file_util import gzip_file, gzip_string, remove_path, tar_gzip_directory, un_gzip_stream, un_gzip_string, un_tar_directory
+from codalabworker.file_util import gzip_file, gzip_string, remove_path, tar_gzip_directory, un_gzip_stream, un_gzip_string, un_tar_directory
 
 
 class FileUtilTest(unittest.TestCase):
